apps/python-fastapi-server/requirements.txt
fastapi==0.115.13
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.25.2
pydantic==2.11.7
aiofiles==23.2.1 